feat(web): add legal pages
